Capture hook to tail call into
Implement a "hook" and utility function that XDP programs can use to
tail call into a PROG_ARRAY_MAP using the action as the index, and then
exit with that action (should the map not contain any program IDs).

Add a go interface for creating the map, rewriting XDP programs to use
it, and pinning it to a bpffs.