Rework flag parsing, add tests
Rework flag / argument parsing to not use the default flag global
FlagSet, as it uses an `ExitOnError` policy which makes it impossible to
test.

Instead create our own FlagSet, and wrap it so that it always
returns proper errors, and doesn't print to stderr on error.
This has the added benefit of better errors, for example if one
of the main arguments is missing.

Write some tests to make sure the flag parsing works as expected.