Capture all actions by default
We currently capture from all known XDP actions by default. This isn't
ideal, as a hook map may expose less actions, leading to an error. More
actions may also be exposed, but won't capture from them by default.

Rework filter so that if filterOpts.actions is empty, we use all the
actions exposed by the hook map. Make this the default in flags.

Rework xdpActions to have a list of all possible actions, so we can use
it in help output. As the default is no longer every action, the names
of actions needs to be discoverable.

Fixes #18