Use a watermark of one by default
Using a higher watermark is in theory more efficient, since the process is woken
up less frequently. This is a problem when using xdpcap in interactive mode however,
because it means that only every 4k of packets are flushed. This makes it look like
no packets were sent when piping to tcpdump.

I suspect that removing the watermark flag completely would probably be fine.