Add automatic loading of hook in XDP frags mode
If an XDP entrypoint is loaded in XDP frags mode, the xdpcap hook
can't be attached in linear buffer mode, and vice versa. This commit
adds a check that allows the code to retry loading the actions in
XDP frags mode if the first attempt fails. If the second attempt fails
too, the program exits and shows the original error message.

A new test is included checking that the attachment retry works as
expected.

Signed-off-by: Pavlos Tzianos <pavlos@cloudflare.com>